# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /library/catalog/enrichment:
    get:
      operationId: getLibraryCatalogEnrichment
      tags: [library]
      summary: Read protected Sheets and Drive library enrichment
      description: Requires an authenticated browser session. Reads the Project Library Sheets tab and live Drive metadata without changing local artifact discovery.
      responses:
        "200":
          description: Protected cloud enrichment with independent source states
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryCatalogEnrichment"
        "401":
          description: Unauthenticated
        "403":
          description: Authenticated user is not authorized for the workspace
  /library/artifacts/{slug}/rebuild:
    post:
      operationId: rebuildLibraryArtifact
      tags: [library]
      summary: Rebuild an approved standalone artifact
      description: Requires an authenticated browser session. Rebuilds only registered artifacts with an approved production command, validates one-file integrity, and atomically hands off the output.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Validated standalone build metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryArtifactBuild"
        "401":
          description: Unauthenticated
        "403":
          description: Authenticated user is not authorized for the workspace
        "404":
          description: Artifact is not registered or cannot be rebuilt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A rebuild could not safely complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Generated output failed standalone integrity validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /library/artifacts/{slug}/export:
    get:
      operationId: downloadLibraryArtifactExport
      tags: [library]
      summary: Download the generated standalone HTML for a project artifact
      description: Requires an authenticated browser session and returns the current generated one-file HTML when the artifact has a production build.
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generated standalone HTML export
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Artifact or generated export was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The source is newer than the generated export or the export failed integrity validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated
...
components:
  schemas:
...
    LibraryArtifactBuild:
      type: object
      required: [state, message]
      properties:
        state:
          type: string
          enum: [current, stale, unavailable]
        message:
          type: string
        builtAt:
          type: ["string", "null"]
          format: date-time
        sourceModifiedAt:
          type: ["string", "null"]
          format: date-time
        size:
          type: ["string", "null"]
        downloadUrl:
          type: ["string", "null"]
    LibraryRecord:
      type: object
      required: [id, title, type, status, description, artifactSlug, artifactUrl, tags, updatedAt, driveFileId, driveWebViewLink]
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        status:
          type: string
        description:
          type: string
        artifactSlug:
          type: ["string", "null"]
        artifactUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        updatedAt:
          type: ["string", "null"]
        driveFileId:
          type: ["string", "null"]
        driveWebViewLink:
          type: ["string", "null"]
    LibrarySourceStatus:
      type: object
      required: [state, message]
      properties:
        state:
          type: string
          enum: [connected, read_only, unavailable, local]
        message:
          type: string
    LibraryDriveFile:
      type: object
      required: [id, name, mimeType, webViewLink, size, modifiedTime, parentIds]
      properties:
        id:
          type: string
        name:
          type: string
        mimeType:
          type: string
        webViewLink:
          type: string
        size:
          type: ["string", "null"]
        modifiedTime:
          type: ["string", "null"]
        parentIds:
          type: array
          items:
            type: string
...
    LibraryCatalogEnrichment:
      type: object
      required: [records, driveFiles, sources]
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/LibraryRecord"
        driveFiles:
          type: array
          items:
            $ref: "#/components/schemas/LibraryDriveFile"
        sources:
          type: object
          required: [sheets, drive]
          properties:
            sheets:
              $ref: "#/components/schemas/LibrarySourceStatus"
            drive:
              $ref: "#/components/schemas/LibrarySourceStatus"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        category:
          type: string
        recoverable:
          type: boolean
...